Fix the color-border-muted variable
It was broken, so the border-based underline below h1 and h2 wasn't
painted, which made the layout shift up 1px, which hampered comparisons
between Markdown Preview Github Styling and this.

I tried to make the variable depend on the editor foreground color
(which we're already using for most of the text in the preview) and add
an alpha channel.

But it's not possible to do that with CSS only [1]. I settled for the
indent guide's color, which is close enough (in the sense of a subtler
version of the default FG color).

If there's really a need to implement the default FG-based version, I
can do that with JS:

- retrieve a theme color with the ThemeColor API [2] (can only be a
    workbench color, syntax colors are not exposed)
- operate on it with any JS color library
- regenerate the colors.css stylesheet with every color variable we need
    (probably doesn't even require a CSS-in-JS library, there's a single
    selector involved, and it's just a mapping of variable names to
    hex/rgba/hsl values)

But I should place a listener on theme changes because this colors
stylesheet will need to be refreshed for every theme.

All in all, too complicated for now. The current setup works with
Lucario and I don't change theme much these days.

ETA: did a quick test and it also works for most of my favorite themes.
So this will stay until I need it (if I make any new themes) or someone
complains about it.

[1] https://stackoverflow.com/questions/40010597/how-do-i-apply-opacity-to-a-css-color-variable
[2] https://code.visualstudio.com/api/references/vscode-api#ThemeColor