Color link :focus outline same as link text
This fixes an extremely off-putting behavior whereby I would click on a
link, and briefly while the browser application opens it in the
background, the link would be outlined in orange.

A flash of unwanted-styled content (FOUC v2), if you will.

This orange outline comes from on VS Code's built-in markdown.css:

a:focus,
... {
  outline: 1px solid -webkit-focus-ring-color;
}

`-webkit-focus-ring-color` is defined as an orange color in Chromium /
Chrome, which I'm assuming sits below VS Code / Electron. [1,2]

But in my home theme (Lucario), a very similar orange serves as
`textPreformat.foreground`, the color for inline code (backticks).

To address this gripe, and because removing Focus Outline is generally
bad, I'm making it the same color as `textLink.foreground` (link blue in
Lucario's case).

This variable rather than `textLink.activeForeground` because the later
comes into play on hover and click (active), but the outline happens
slightly after the click. Which has a reasonable chance of being :hover...
Whatever, let's roll with this cause it's 10:40pm. I wanna go home.

[1] https://stackoverflow.com/questions/7538771/what-is-webkit-focus-ring-color
[2] https://www.color-hex.com/color/e59700