Make the color for <hr> and table borders more neutral
Currently Lucario Theme uses #666 for `editorIndentGuide.background`,
which looks groovy on it's background blue.

By changing our `--color-border-default` to `tab.unfocusedActiveForeground`,
we're making it more neutral (perceived as white against the blue
backdrop), and also more similar to the theme's foreground color.

Other Lucario color settings that use this anchor (&fg-subtle) are:

activityBar.inactiveForeground
tab.unfocusedInactiveForeground
gitDecoration.ignoredResourceForeground
...

We may want to switch to one of those if it covers the average theme
better than this.

And we may want to switch back to `editorIndentGuide` colors (`.background`
and `.activeBackground`) if I they're fixed on Lucario Theme.

Also, CSS vars can use fallbacks, so I may define several vscode color
variables to increase the chance that one is a hit.