Color `inline code` fg like default foreground
It was the same color as `inline code` in the .md source before.

Has been a bit of a pain point since the textPreformat color for my
Lucario Theme is orange, which kinda breaks the flow of text. Also cause
I use lots of `inline code` in my notes.

Also darkens the background color for those inline code spans, so it
becomes more visible, cause it lacked contrast with normal text, after
changing the orange foreground.